backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
pydantic
google-api-python-client
requests
EOL